requirements.txt
python-dotenv==1.0.1
pytest-playwright==0.5.2
